Correct project layout for gemification
The original project structure was never intended to be gemified. These
changes structure the project internally to better suit that, and
include a gem specification file.

For the timebeing, wraith is being loaded from the `Gemfile` since the
HTML differ requires our fork, which adds configuration for concurrency.

I'll document this separately.